+++ pages/_app.js
@@ -1,8 +1,6 @@
 import '@/styles/globals.css';
 import localFont from 'next/font/local';
 import AppShell from '@/components/AppShell';
-import SkmPrompt from '@/components/SkmPrompt';
-import { PUBLIK_AKTIF } from '@/lib/modeSitus';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
 import { Analytics } from '@vercel/analytics/react';
@@ -67,7 +65,6 @@ export default function App({ Component, pageProps }) {
         <AppShell>
           <Component {...pageProps} />
         </AppShell>
-        {PUBLIK_AKTIF && <SkmPrompt />}
       </div>
       <Analytics />
     </>